Move trigger trie to own file
